fix(commands): add top-level docs command
