+++ src/types/command/commandFactory.ts
@@ -12,9 +12,7 @@ import { CreateCommand } from '../../commands/create/createCommand'
 import { DbCommand } from '../../commands/db/dbCommand'
 import { CompileBuildDeployCommand } from '../../commands/deploy/compileBuildDeployCommand'
 import { DeployCommand } from '../../commands/deploy/deployCommand'
-import { GenerateDocsCommand } from '../../commands/docs/generateDocsCommand'
-import { GenerateDotCommand } from '../../commands/docs/generateDotCommand'
-import { InitDocsCommand } from '../../commands/docs/initDocsCommand'
+import { DocsCommand } from '../../commands/docs/docsCommand'
 import { TestCommand } from '../../commands/testing/testCommand'
 import { ListFolderCommand } from '../../commands/folder/listFolderCommand'
 import { VersionCommand } from '../../commands/version/versionCommand'
@@ -42,9 +40,7 @@ export const commandFactory = new Map<string, Function>([
   ['context list', (args: string[]) => new ListContextCommand(args)],
   ['db', (args: string[]) => new DbCommand(args)],
   ['deploy', (args: string[]) => new DeployCommand(args)],
-  ['doc', (args: string[]) => new GenerateDocsCommand(args)],
-  ['doc init', (args: string[]) => new InitDocsCommand(args)],
-  ['doc lineage', (args: string[]) => new GenerateDotCommand(args)],
+  ['doc', (args: string[]) => new DocsCommand(args)],
   ['folder create', (args: string[]) => new CreateFolderCommand(args)],
   ['folder delete', (args: string[]) => new DeleteFolderCommand(args)],
   ['folder list', (args: string[]) => new ListFolderCommand(args)],
